ci: specialize config check CI workflow
Update the CI workflow for validating the Codecov configuration from
being framed as a "general config check workflow" to a specific "Codecov
config check workflow". The main aim of this change is to avoid, in the
future, running the Codecov configuration check when another config
file that is continuously validated is changed. This wastes resources
and puts unnecessary load on the Codecov API.