chore!: drop support for SVGO v1
End support for SVGO v1 entirely for the project after dropping built-in
support in major version 3 of this Action. SVGO v1 has been deprecated
for a long time and really shouldn't be used anymore. Removing support
has quite an impact in terms of simplification of the code base of this
Action.