docs: use `actions/checkout` v4 in example workflows
Update all example workflows using `actions/checkout` in the docs to use
v4 instead of v3. This is to encourage users of this action to also use
the latest version of other actions.

As v4 of the action runs on a later Node.js runtime than this action,
this change is intended to be included when this action is also bumped
to the latest Node.js runtime.