feat!: replace `minimatch` by `(at)actions/glob`
Switch to `(at)actions/glob` in an effort to align globbing for this
Action with the larger Actions ecosystem (assuming most actions will use
the officially provided globbing library).

The aim of the implementation seen here is primarily to perform the
replacement. The API of `(at)actions/glob` does not lend itself well to
how it is being used here - future work could be to refactor the
internals of this action to better align with the `(at)actions/glob`
API.